+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Added public function "getId()", which will return the element ID of the calendar instance.
 - Added back in version stamp into the minimized version of the main JS file.
 - Added new custom event "onDestroy", which is triggered when a Calendar instance is destroyed (the Calendar ID is passed).
+- Added new custom event "onRefresh", which is triggered when the Calendar is manually refreshed.
 - 
 
 ### Version 1.6.6: